renamed files
renamed all the sub devices to a common name. Moved all the devices in its own folder.
created Utils and Tests folders
