--- broadlink/setup.py
@@ -1,9 +1,0 @@
-#!/usr/bin/python
-
-from __init__ import setup, discover
-
-# -----------------------------------------------------------------
-# initi program: 
-
-print "Entering in setup mode"
-setup('Chuck_Norris', '8ruceL33+', 3)
