+++ broadlink/tests/sp2mini2_test.py
@@ -3,22 +3,26 @@
 from __init__ import setup, discover
 import time
 
-# -----------------------------------------------------------------
-# initi program: 
+# Notice:
+# The divices have to be pre-configured and associated 
+# to the local Wlan before running this test
+
+# usage:
+# python sp2mini2_test.py
 
 print "Call discovery..."
 devices = discover(timeout=5)
 print "Found %d" % len(devices)
-#print "device SPMini2: Sp2 class"
 
 if len(devices) > 0:
-	#if devices[0].get_type == "SP2Mini2":
 	print "device type %s" % devices[0].get_type
-	#print "device SPMini2: Sp2 class"
+
 	print "Call Auth"
 	devices[0].auth()
+
 	print "Call set power"
 	devices[0].set_power(1)
+
 	print "Call check power"
 	devices[0].check_power()
 
@@ -27,5 +31,6 @@ if len(devices) > 0:
 	
 	print "Call set power"
 	devices[0].set_power(0)
+	
 	print "Call check power"
 	devices[0].check_power()
